+++ internal/domain/item/service.go
@@ -31,14 +31,14 @@ func (p *ItemService) Create(item entities.Item) error {
 	return p.itemRepository.Create(ctx, item)
 }
 
-func (p *ItemService) GetById(itemID int64) (entities.Item, error) {
+func (p *ItemService) GetById(itemID int) (entities.Item, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
 	defer cancel()
 
 	return p.itemRepository.GetById(ctx, itemID)
 }
 
-func (p *ItemService) Delete(itemID int64) error {
+func (p *ItemService) Delete(itemID int) error {
 	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
 	defer cancel()
